Brings in ChunkReviewDao::lockByJobId(), IDatabase::onCommit() and the cache busts that now run past
the commit rather than inside the transaction that wrote the row.

Three adaptations were needed where the branch predates conventions this one introduced: EventModel's
test reset $__transactionStarted as a static, which is per-object here; two of its transaction tests
used a connection stub that always reports no transaction, which the stricter commit and rollback
guards read as nothing to close; and the drift alert template escaped by hand, which double-encodes
now that EmailValue wraps nested rows.

internal_scripts moves to a commit that merges master into the drift branch, so the tasks and the
team name decoder are both present whichever repository lands first.

+++ lib/Plugins/Features/ReviewExtended/ReviewedWordCountModel.php
@@ -432,18 +432,22 @@ class ReviewedWordCountModel implements IReviewedWordCountModel
      *
      * @param int $source_page
      *
-     * @return int
+     * EntryStruct::$penalty_points is a float, so the sum must be too — an int return type here
+     * truncates (two issues of 2.75 subtract 5 instead of 5.50) and the difference is left behind
+     * on the chunk review row as drift.
+     *
+     * @return float
      */
-    private function getPenaltyPointsForSourcePage(int $source_page): int
+    private function getPenaltyPointsForSourcePage(int $source_page): float
     {
         $toReduce = $this->_event->getIssuesToDelete();
         $issues = array_filter($toReduce, function (EntryStruct $issue) use ($source_page) {
             return $issue->source_page == $source_page;
         });
 
-        return array_reduce($issues, function ($carry, EntryStruct $issue) {
+        return (float)array_reduce($issues, function ($carry, EntryStruct $issue) {
             return $carry + $issue->penalty_points;
-        }, 0);
+        }, 0.0);
     }
 
 }
